agent/requirements.txt
requests==2.32.3
pyyaml==6.0.2
psutil==5.9.8

